requirements.txt
Flask
Flask-SocketIO
paho-mqtt
Werkzeug